requirements.txt
Flask==3.0.0
pytest==7.0.1; python_version > "3.0"
# pin pytest to 4.6.11 for Python2.
pytest==7.0.1; python_version < "3.0"
gunicorn==20.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
distro-info==1.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-cloud-resource-manager==1.5.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-storage==2.9.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
packaging==23.1
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
werkzeug>=3.0.0